Create an AuthCode PORO
Created an AuthCode Ruby object to wrap around our JWT library. This
object is responsible for encoding and decoding user information such
as username, id and email.

We then return that in json after receiving a request to the signin
endpoint. Once we have that token we can use it to authenticate our
requests.

There is an example of how to use fetch once you have your token.

We also added the model of Deals just to have a resource
on the backend that will require Authentication and Authorization to
CRUD.